project/brand_drf/requirements.txt
Django==2.0.7
djangorestframework==3.8.2
whitenoise==3.3.1
